Fix function reference rerender in BottomTabBar
