+++ packages/mobile/src/components/bottom-tab-bar/index.ts
@@ -1,5 +1,5 @@
 import { BottomTabBarProps as _BottomTabBarProps } from './BottomTabBar'
 
-export { default } from './BottomTabBar'
+export { BottomTabBar } from './BottomTabBar'
 
 export type BottomTabBarProps = _BottomTabBarProps
